+++ source/ApplyDBMigrationsApp/Helpers/ConnectionStringParser.cs
@@ -12,34 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using DbUp.Engine;
-
 namespace Energinet.DataHub.EDI.ApplyDBMigrationsApp.Helpers;
 
-public static class DbUpgradeRunner
+internal static class ConnectionStringParser
 {
-    private static bool _isRunning;
-
-    public static DatabaseUpgradeResult RunDbUpgrade(
-        string connectionString,
-        string environment = "",
-        bool isDryRun = false)
+    public static string Parse(string[] args)
     {
-        while (_isRunning)
-        {
-            //To avoid both database fixtures from performing the upgrade at the same time
-            Thread.Sleep(2000);
-        }
-
-        _isRunning = true;
-        var upgrader = UpgradeFactory.GetUpgradeEngine(
-            connectionString,
-            environment,
-            isDryRun);
-
-        var result = upgrader.PerformUpgrade();
-
-        _isRunning = false;
-        return result;
+        return args.First();
     }
 }
