+++ source/ApplyDBMigrationsApp/Helpers/ResultReporter.cs
@@ -16,7 +16,7 @@ using DbUp.Engine;
 
 namespace Energinet.DataHub.EDI.ApplyDBMigrationsApp.Helpers;
 
-public static class ResultReporter
+internal static class ResultReporter
 {
     private const int SuccessResult = 0;
     private const int FailureResult = -1;
